requirements.txt
streamlit-authenticator
streamlit-cropper
streamlit-webrtc